✨ feat: add workflow parser for trophy card generation
